made a src folder and connected sql
