+++ .gitignore
@@ -2,3 +2,4 @@
 /echo.class
 /main$Book.class
 /main$Administrator.class
+/bin/
