+++ classPackage/.gitignore
@@ -14,3 +14,5 @@
 /DateNumber.class
 /Wishlist.class
 /UserObject.class
+/Main$Administrator.class
+/Main.class
