--- mainPackage/main.java
@@ -1,36 +1,0 @@
-package mainPackage;
-//import java.util.*;
-
-public class main{
-
-
-    class membership{
-
-        String Fname;
-        String Lname;
-        String Minitial;
-        String email;
-        int number;
-        String address;
-        String password;
-        String username;
-        String dob;
-        boolean isMember;
-        int dues;
-
-    }
-
-    class Administrator{
-        String role;
-        String name;
-        String email;
-        int adminid;
-    }
-
-    public static void main(String[] args){
-
-        System.out.println("Hello");
-        System.out.println("Hello World!");
-
-    }
-}
